refactor: human input module
