+++ packages/server/api/src/app/flows/flow/human-input/human-input.module.ts
@@ -1,36 +1,8 @@
-import {
-    ALL_PRINCIPAL_TYPES,
-    ApId,
-    USE_DRAFT_QUERY_PARAM_NAME,
-} from '@activepieces/shared'
-import { FastifyPluginAsyncTypebox, Type } from '@fastify/type-provider-typebox'
-import { humanInputService } from './human-input.service'
-
+import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox'
+import { formController } from './form-controller'
+import { chatController } from './chat-controller'
 
 export const humanInputModule: FastifyPluginAsyncTypebox = async (app) => {
-    await app.register(humanInputController, { prefix: '/v1/human-input' })
-}
-
-const humanInputController: FastifyPluginAsyncTypebox = async (app) => {
-    app.get('/form/:flowId', GetFormRequest, async (request) => {
-        return humanInputService(request.log).getFormByFlowIdOrThrow(request.params.flowId, request.query.useDraft ?? false)
-    })
-    app.get('/chat/:flowId', GetFormRequest, async (request) => {
-        return humanInputService(request.log).getChatUIByFlowIdOrThrow(request.params.flowId, request.query.useDraft ?? false)
-    })
-}
-
-const GetFormRequest = {
-    config: {
-        allowedPrincipals: ALL_PRINCIPAL_TYPES,
-    },
-    schema: {
-        description: 'Get a form by flow id',
-        params: Type.Object({
-            flowId: ApId,
-        }),
-        querystring: Type.Object({
-            [USE_DRAFT_QUERY_PARAM_NAME]: Type.Optional(Type.Boolean()),
-        }),
-    },
+    await app.register(formController, { prefix: '/v1/human-input' })
+    await app.register(chatController, { prefix: '/v1/human-input' })
 }
